refactor(factoryconfig): split worker and workstation mapping
